initialize/check lib/context in CommUCC (#45)
Summary:
Pull Request resolved: https://github.com/facebookresearch/torch_ucc/pull/45

Noticed that the tcpStore can be gone before the ProcessGroup is destroyed (destructor invoked) during GC.
Avoid getting stuck or get into abort when tcpStore is gone before the completion of the ProcessGroup.
This is especially true when a oob_allgather is needed in the exit path (which relies on tcpStore) and the tcpStore is gone.

Also : Hardening - set/check for null for critical handles in the exit path

Reviewed By: kingchc, srinivas212

Differential Revision: D32026919

fbshipit-source-id: 251939275161060b7b5c89c137e9cdf285e44ac4