Use new UCC pt2pt calls
Summary:
Rather than calling UCX directly from TORCH-UCC to make point-to-point send and
receive calls, call the new interfaces in UCC to be more portable and take
advantage of NCCL when possible.

This does have trade-offs of removing support for an any-source receive, which
seems to have been unsupported in NCCL-PG regardless. This also removes support
for tagged messages as NCCL does not support them.

Reviewed By: minsii

Differential Revision: D35432333

fbshipit-source-id: 5b0e225c73b36571113fb12335de1e3e920e8909