Run PyTorch unit tests in OSS CI (#69)
Summary:
The `torch_tests.py` was taken from https://github.com/pytorch/pytorch/pull/70049, and I modified the skip list. I am running these tests in our NVIDIA CI, and I find that changes in `torch.testing._internal.distributed.distributed_test` might break this test. I think it is a good idea to maintain a copy of this file in the public `torch_ucc` repository so that we can get the signal of the status of these tests from the public OSS CI.

Pull Request resolved: https://github.com/facebookresearch/torch_ucc/pull/69

Reviewed By: kingchc

Differential Revision: D35897723

Pulled By: bryanmr

fbshipit-source-id: 4225ba4184bc9bfd6a34a1bdf62db92f26630050