completely disable ucx when ACTIVE_SET is on
Summary:
When UCC provides ACTIVE_SET feature, we can completely disable all ucx usage in torch_ucc and remove ucp and ucs dependencies.

Limitation of this commit:
The following API is not yet supported in the ACTIVE_SET version, thus simply report error.  It is the same behavior in NCCL PG: recvAnysource

Reviewed By: wesbland, kingchc

Differential Revision: D38142280

fbshipit-source-id: bfc8b9d804e73c5237f9011311dd756e1fe8c6e9